♻️ Refactor: 자잘한 수정
