+++ src/app/mypage/_components/AppointmentCard.tsx
@@ -1,15 +1,20 @@
-import { useProfileImageQuery } from '@/hooks/queries/useProfileImage';
+import { useProfileImageQuery } from '@/hooks/queries/useUserQuery';
+import { MyAppointment } from '@/types/appointments';
+import { ImageType } from '@/types/image';
 import { Avatar, AvatarFallback, AvatarImage } from '@/ui/shadcn/avatar';
 import { Button } from '@/ui/shadcn/button';
 
-const AppointmentCard = ({ data }) => {
+const AppointmentCard = ({ data }: { data: MyAppointment }) => {
   const { appointments, other_user_nickname, other_user_profile } = data;
   const { data: image } = useProfileImageQuery(other_user_profile, {
     enabled: !!other_user_profile,
   });
+
+  const imageUrl = image as ImageType;
+
   const profileImage = other_user_profile.includes('found_default')
     ? other_user_profile
-    : image?.publicUrl;
+    : imageUrl?.publicUrl;
 
   const appointment = appointments[0];
