--- src/app/mypage/page.tsx
@@ -1,13 +1,0 @@
-import UserProfile from '@/app/mypage/_components/UserProfile';
-import Appointments from './_components/Appointments';
-
-const MyPage = () => {
-  return (
-    <div className="flex justify-center gap-20">
-      <UserProfile />
-      <Appointments />
-    </div>
-  );
-};
-
-export default MyPage;
