+++ src/hooks/mutations/useEditProfileMutation.ts
@@ -1,3 +1,4 @@
+import { QUERY_KEY } from '@/constants/constants';
 import { editProfile } from '@/services/profileServices';
 import { UserData } from '@/types/users';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
@@ -8,7 +9,9 @@ export const useEditProfileMutation = (user_id: number) => {
   return useMutation({
     mutationFn: ({ data }: { data: UserData }) => editProfile(user_id, data),
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: [`userData-${user_id}`] });
+      queryClient.invalidateQueries({
+        queryKey: [QUERY_KEY.USERS, user_id],
+      });
     },
     onError: (error) => {
       console.error('프로필 수정 에러', error);
