+++ src/hooks/mutations/useUploadImageMutation.ts
@@ -1,5 +1,6 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { uploadImage } from '../../services/uploadProfile';
+import { QUERY_KEY } from '@/constants/constants';
 
 export const useUploadProfileImage = (user_id: number) => {
   const queryClient = useQueryClient();
@@ -8,7 +9,9 @@ export const useUploadProfileImage = (user_id: number) => {
     mutationFn: ({ filePath, file }: { filePath: string; file: File }) =>
       uploadImage(filePath, file),
     onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: [`userData-${user_id}`] });
+      queryClient.invalidateQueries({
+        queryKey: [QUERY_KEY.USERS, user_id],
+      });
     },
     onError: (error: Error) => {
       console.error('Image upload error', error.message);
