+++ src/types/appointments.ts
@@ -14,3 +14,13 @@ export type AppointmentInputs = {
   date: string;
   category: string[];
 };
+
+export type MyAppointment = {
+  appointments: Appointment[];
+  id: number;
+  otherUserId: number;
+  other_user_nickname: string;
+  other_user_profile: string;
+  user1_id: number;
+  user2_id: number;
+};
